refactor(desktop): migrate presentation package to desktop
migrate all resources, controllers, FXML files, and stylesheets from the presentation package to the desktop package to improve module organization and clarify intent. update imports, file paths, and references across the codebase accordingly.
