+++ src/main/java/br/com/gzlabs/gzassist/application/AppFactory.java
@@ -5,8 +5,8 @@ import br.com.gzlabs.gzassist.adapters.OpenAiAnswerProvider;
 import br.com.gzlabs.gzassist.adapters.RobotScreenCapturer;
 import br.com.gzlabs.gzassist.config.AppConfig;
 import br.com.gzlabs.gzassist.core.Mode;
+import br.com.gzlabs.gzassist.desktop.UiEvent;
 import br.com.gzlabs.gzassist.errors.HotkeyException;
-import br.com.gzlabs.gzassist.presentation.UiEvent;
 import br.com.gzlabs.gzassist.util.PromptTemplates;
 import com.openai.client.OpenAIClient;
 import com.openai.client.okhttp.OpenAIOkHttpClient;
