+++ src/main/java/br/com/gzlabs/gzassist/util/ThemeManager.java
@@ -21,8 +21,8 @@ public class ThemeManager {
     private static void apply(Scene scene, Parent root, String themeFile, String themeClass) {
         scene.getStylesheets().clear();
         scene.getStylesheets().addAll(
-                Objects.requireNonNull(ThemeManager.class.getResource("/br/com/gzlabs/gzassist/presentation/style.css")).toExternalForm(),
-                Objects.requireNonNull(ThemeManager.class.getResource("/br/com/gzlabs/gzassist/presentation/" + themeFile)).toExternalForm()
+                Objects.requireNonNull(ThemeManager.class.getResource("/desktop/home/style.css")).toExternalForm(),
+                Objects.requireNonNull(ThemeManager.class.getResource("/desktop/theme/" + themeFile)).toExternalForm()
         );
         root.getStyleClass().removeAll("theme-dark", "theme-light");
         root.getStyleClass().add(themeClass);
