+++ src/main/java/module-info.java
@@ -9,10 +9,10 @@ module br.com.gzlabs.gzassist {
     requires java.prefs;
 
     exports br.com.gzlabs.gzassist.core;
-    exports br.com.gzlabs.gzassist.presentation;
+    exports br.com.gzlabs.gzassist.desktop;
     exports br.com.gzlabs.gzassist.application;
     exports br.com.gzlabs.gzassist.errors;
     exports br.com.gzlabs.gzassist;
 
-    opens br.com.gzlabs.gzassist.presentation to javafx.fxml;
+    opens br.com.gzlabs.gzassist.desktop to javafx.fxml;
 }
