+++ src/main/resources/desktop/theme/base.css
