+++ src/main/resources/desktop/theme/theme-light.css
